feat: add ts migration
