+++ package.json
@@ -5,11 +5,10 @@
     "type": "git",
     "url": ""
   },
-  "type": "commonjs",
   "description": "nodejs-mitso",
-  "main": "src/server.js",
+  "main": "src/server.ts",
   "scripts": {
-    "start": "nodemon src/server.js",
+    "start": "nodemon src/server.ts",
     "lint": "eslint --ext .js,.ts --fix src/"
   },
   "keywords": [
@@ -44,11 +43,13 @@
     "dotenv": "9.0.0",
     "express": "4.17.1",
     "http-status-codes": "^2.1.4",
-    "uuid": "8.3.2"
+    "swagger-ui-express": "^4.2.0",
+    "uuid": "8.3.2",
+    "yamljs": "^0.3.0"
   },
   "devDependencies": {
     "debug": "4.3.1",
-    "eslint": "^8.4.1",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "14.2.1",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-import": "2.22.1",
