+++ src/server.ts
@@ -1,7 +1,5 @@
-const {
-  PORT
-} = require('./common/config');
-const app = require('./app');
+import { PORT } from './common/config';
+import app from './app';
 
 app.listen(PORT, () =>
   console.log(`App is running on http://localhost:${PORT}`)
